add raw code to feature type files
