+++ requirements.txt
@@ -1,0 +1,6 @@
+pandas
+numpy
+pyspark
+matplotlib
+xgboost
+pandasql
